rewording suggestions
I proofread the text for clarity and readability. I made a few suggested changes, but also have the following questions:

In line 56 --is it correct to say "An attestation is an authenticated metadata" as opposed to a 
"piece of metadata"? It feels like there is an agreement issue here.

In line 83, is "inspecting" the right word? 

In line 87, "must use to parse the contents of a predicate," perhaps?