ITE-9: Require predicateName for predicates
The short predicateName is expected to be unique as well, and is to be
used for things like the file name for the predicate as well as special
media types that also indicate the predicate contained within.

In addition, this commit updates a link to the in-toto attestation spec,
pointing to the formal v1.0 release instead of the older draft version.

Signed-off-by: Aditya Sirish <aditya@saky.in>
Co-authored-by: Marcela Melara <marcela.melara@intel.com>